+++ src/components/Widget/components/Conversation/components/Header/index.js
@@ -4,23 +4,33 @@ import PropTypes from 'prop-types';
 import close from 'assets/clear-button.svg';
 import './style.scss';
 
-const Header = ({ title, subtitle, toggleChat, showCloseButton }) =>
-  <div className="header">
+const Header = ({ title, subtitle, toggleChat, showCloseButton, connected, connectingText }) =>
+  <div>
+    <div className="header">
     {
       showCloseButton &&
       <button className="close-button" onClick={toggleChat}>
         <img src={close} className="close" alt="close" />
       </button>
     }
-    <h4 className="title">{title}</h4>
-    {subtitle && <span>{subtitle}</span>}
-  </div>;
+      <h4 className="title">{title}</h4>
+      {subtitle && <span>{subtitle}</span>}
+    </div>
+  {
+    !connected &&
+    <span className="loading">
+      {connectingText}
+    </span>
+  }
+  </div>
 
 Header.propTypes = {
   title: PropTypes.string,
   subtitle: PropTypes.string,
   toggleChat: PropTypes.func,
-  showCloseButton: PropTypes.bool
+  showCloseButton: PropTypes.bool,
+  connected: PropTypes.bool,
+  connectingText: PropTypes.string
 };
 
 export default Header;