+++ src/components/Widget/components/Conversation/components/Messages/components/index.js
@@ -1,0 +1,13 @@
+import Video from './VidReply';
+import Image from './ImgReply';
+import Message from './Message';
+import Snippet from './Snippet';
+import QuickReply from './QuickReply';
+
+export {
+    Video,
+    Image,
+    Message,
+    Snippet,
+    QuickReply
+}