+++ src/components/Widget/components/Launcher/index.js
@@ -1,17 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
 import Badge from './components/Badge';
 
 import openLauncher from 'assets/launcher_button.svg';
 import close from 'assets/clear-button.svg';
 import './style.scss';
 
-const Launcher = ({ toggle, chatOpened, badge }) =>
-  <button type="button" className={chatOpened ? 'launcher hide-sm' : 'launcher'} onClick={toggle}>
+const Launcher = ({ toggle, isChatOpen, badge, fullScreenMode }) =>
+  <button type="button" className={(isChatOpen ? 'launcher hide-sm' : 'launcher') + (fullScreenMode ? ` full-screen${isChatOpen ? '  hide' : ''}` : '')} onClick={toggle}>
     <Badge badge ={badge} />
     {
-      chatOpened ?
+      isChatOpen ?
         <img src={close} className="close-launcher" alt="" /> :
         <img src={openLauncher} className="open-launcher" alt="" />
     }
@@ -19,10 +18,9 @@ const Launcher = ({ toggle, chatOpened, badge }) =>
 
 Launcher.propTypes = {
   toggle: PropTypes.func,
-  chatOpened: PropTypes.bool,
-  badge: PropTypes.number
+  isChatOpen: PropTypes.bool,
+  badge: PropTypes.number,
+  fullScreenMode: PropTypes.bool
 };
 
-export default connect(store => ({
-  chatOpened: store.behavior.get('showChat')
-}))(Launcher);
+export default Launcher;