+++ src/store/reducers/helper.js
@@ -1,11 +1,8 @@
 import { Map, List } from 'immutable';
-import { MESSAGES_TYPES, MESSAGE_SENDER } from 'constants';
+import { MESSAGES_TYPES, MESSAGE_SENDER, SESSION_NAME } from 'constants';
 
-import Video from 'messagesComponents/VidReply';
-import Image from 'messagesComponents/ImgReply';
-import Message from 'messagesComponents/Message';
-import Snippet from 'messagesComponents/Snippet';
-import QuickReply from 'messagesComponents/QuickReply';
+import { Video, Image, Message, Snippet, QuickReply } from 'messagesComponents';
+
 
 export function createNewMessage(text, sender) {
   return Map({
@@ -73,4 +70,75 @@ export function createComponentMessage(component, props, showAvatar) {
     sender: MESSAGE_SENDER.RESPONSE,
     showAvatar
   });
+}
+
+export function getLocalSession(storage, key) {
+  // Attempt to get local session from storage
+  const cachedSession = storage.getItem(key);
+  var session = null;
+  if (cachedSession) {
+    // Found existing session in storage
+    let parsedSession = JSON.parse(cachedSession)
+    // Format conversation from array of object to immutable Map for use by messages components
+    const formattedConversation = parsedSession.conversation
+      ? Object.values(parsedSession.conversation).map(item => Map(item))
+      : [];
+    // Check if params is undefined
+    const formattedParams = parsedSession.params
+      ? parsedSession.params
+      : {};
+    //Create a new session to return
+      session = {
+      ...parsedSession,
+      conversation: formattedConversation,
+      params: formattedParams
+    }
+  }
+  // Returns a formatted session object if any found, otherwise return undefined
+  return session;
+}
+
+export function storeLocalSession(storage, key, sid) {
+  // Attempt to store session id to local storage
+  const cachedSession = storage.getItem(key);
+  var session;
+  if (cachedSession) {
+      // Found exisiting session in storage
+    let parsedSession = JSON.parse(cachedSession)
+    session = {
+      ...parsedSession,
+      session_id: sid
+    }
+  } else {
+    // No existing local session, create a new empty session with only session_id
+    session = {
+      session_id: sid
+    }
+  }
+  // Store updated session to storage
+  storage.setItem(key, JSON.stringify(session));
+}
+
+export const storeMessageTo = (storage) => (conversation) => {
+  // Store a conversation List to storage
+  const localSession = getLocalSession(storage, SESSION_NAME);
+  const newSession = {
+    // Since immutable List is not a native JS object, store conversation as array
+    ...localSession,
+    conversation: [...Array.from(conversation)]
+  }
+  storage.setItem(SESSION_NAME, JSON.stringify(newSession));
+  return conversation
+}
+
+export const storeParamsTo = (storage) => (params) => {
+  // Store a params List to storage
+  const localSession = getLocalSession(storage, SESSION_NAME);
+  const newSession = {
+    // Since immutable Map is not a native JS object, store conversation as array
+    ...localSession,
+    params: params.toJS()
+  }
+  storage.setItem(SESSION_NAME, JSON.stringify(newSession));
+  return params
 }