+++ src/store/store.js
@@ -1,29 +1,37 @@
 import { createStore, combineReducers, applyMiddleware } from "redux";
 
+import { SESSION_NAME } from 'constants';
+
 import behavior from "./reducers/behaviorReducer";
 import messages from "./reducers/messagesReducer";
+import { getLocalSession } from './reducers/helper';
 import * as actionTypes from './actions/actionTypes';
 
 let store = "call initStore first";
 
-function initStore(hint, socket) {
+function initStore(hintText, connectingText, socket, storage) {
+
   const customMiddleWare = (store) => next => (action) => {
+    const session_id = (getLocalSession(storage, SESSION_NAME)? getLocalSession(storage, SESSION_NAME).session_id: null);
     switch (action.type) {
       case actionTypes.EMIT_NEW_USER_MESSAGE: {
-        socket.emit("user_uttered", { message: action.text, customData: socket.customData });
+        socket.emit("user_uttered", { message: action.text, customData: socket.customData, session_id });
       }
       case actionTypes.GET_OPEN_STATE: {
-        return store.getState().behavior.get("showChat");
+        return store.getState().behavior.get("isChatOpen");
       }
       case actionTypes.GET_VISIBLE_STATE: {
-        return store.getState().behavior.get("showWidget");
+        return store.getState().behavior.get("isChatVisible");
       }
     }
 
     // console.log('Middleware triggered:', action);
     next(action);
   };
-  const reducer = combineReducers({ behavior: behavior(hint), messages });
+  const reducer = combineReducers({ 
+    behavior: behavior(hintText, connectingText, storage), 
+    messages: messages(storage) 
+  });
 
   /* eslint-disable no-underscore-dangle */
   store = createStore(