+++ webpack.prod.js
@@ -4,7 +4,7 @@ const path = require('path');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
 
 module.exports = {
-  entry: './index.js',
+  entry: ['babel-polyfill', './index.js'],
   output: {
     path: path.join(__dirname, '/lib'),
     filename: 'index.js',